requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0        # async SQLite driver
# psycopg2-binary==2.9.9  # PostgreSQL 備用，目前不使用

# ── Validation & Settings ─────────────────────────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2

# ── LLM ──────────────────────────────────────────────────────────────────────
openai==1.51.0
httpx==0.27.2            # openai SDK 與 httpx>=0.28 不相容，固定此版本

# ── Task Queue (生產模式才需要，DEV_MODE=True 時可不裝) ───────────────────────
# celery==5.4.0
# redis==5.1.1

# ── File Parsing ──────────────────────────────────────────────────────────────
python-docx==1.1.2
pdfplumber==0.11.4

# ── Misc ──────────────────────────────────────────────────────────────────────
python-multipart==0.0.12  # FastAPI 上傳表單
aiofiles==24.1.0          # FastAPI StaticFiles 需要
